ロジックをViewから分離
